Extract non-UI job to GitMacheteRepositoryUpdateTask
