+++ frontendUiApi/src/main/java/com/virtuslab/gitmachete/frontend/ui/api/table/IGraphTable.java
@@ -1,6 +1,12 @@
 package com.virtuslab.gitmachete.frontend.ui.api.table;
 
+import java.nio.file.Path;
+
 import org.checkerframework.checker.guieffect.qual.UIEffect;
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+import com.virtuslab.branchlayout.api.manager.IBranchLayoutWriter;
+import com.virtuslab.gitmachete.backend.api.IGitMacheteRepository;
 
 public interface IGraphTable {
   @UIEffect
@@ -11,4 +17,9 @@ public interface IGraphTable {
 
   @UIEffect
   void refreshModel();
+
+  @UIEffect
+  void refreshModel(@Nullable IGitMacheteRepository gitMacheteRepository, Path macheteFilePath, boolean isMacheteFilePresent);
+
+  void setBranchLayoutWriter(IBranchLayoutWriter branchLayoutWriter);
 }
