+++ frontendUiTableImpl/src/main/java/com/virtuslab/gitmachete/frontend/ui/impl/table/GitMacheteRepositoryUpdateTask.java
@@ -1,0 +1,135 @@
+package com.virtuslab.gitmachete.frontend.ui.impl.table;
+
+import static com.intellij.openapi.application.ModalityState.NON_MODAL;
+import static com.virtuslab.gitmachete.frontend.ui.impl.table.GitPathUtils.getGitDirectoryPath;
+import static com.virtuslab.gitmachete.frontend.ui.impl.table.GitPathUtils.getMacheteFilePath;
+import static com.virtuslab.gitmachete.frontend.ui.impl.table.GitPathUtils.getMainDirectoryPath;
+
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+import com.intellij.openapi.progress.ProgressIndicator;
+import com.intellij.openapi.progress.Task;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.vcs.VcsNotifier;
+import com.intellij.ui.GuiUtils;
+import git4idea.repo.GitRepository;
+import io.vavr.control.Option;
+import io.vavr.control.Try;
+import org.checkerframework.checker.index.qual.Positive;
+
+import com.virtuslab.binding.RuntimeBinding;
+import com.virtuslab.branchlayout.api.BranchLayoutException;
+import com.virtuslab.branchlayout.api.IBranchLayout;
+import com.virtuslab.branchlayout.api.manager.IBranchLayoutManager;
+import com.virtuslab.branchlayout.api.manager.IBranchLayoutManagerFactory;
+import com.virtuslab.gitmachete.backend.api.IGitMacheteRepository;
+import com.virtuslab.gitmachete.backend.api.IGitMacheteRepositoryFactory;
+import com.virtuslab.gitmachete.backend.api.MacheteFileReaderException;
+import com.virtuslab.gitmachete.frontend.ui.api.table.IGraphTable;
+import com.virtuslab.logger.IPrefixedLambdaLogger;
+import com.virtuslab.logger.IntelliJLoggingUtils;
+import com.virtuslab.logger.PrefixedLambdaLoggerFactory;
+
+public final class GitMacheteRepositoryUpdateTask extends Task.Backgroundable {
+
+  private static final IPrefixedLambdaLogger LOG = PrefixedLambdaLoggerFactory.getLogger("frontendUiTable");
+
+  private final Project project;
+  private final GitRepository gitRepository;
+  private final IGraphTable graphTable;
+
+  private final IGitMacheteRepositoryFactory gitMacheteRepositoryFactory;
+  private final IBranchLayoutManagerFactory branchLayoutManagerFactory;
+
+  private GitMacheteRepositoryUpdateTask(Project project, GitRepository gitRepository, IGraphTable graphTable) {
+    super(project, "Updating Git Machete repository");
+
+    this.project = project;
+    this.gitRepository = gitRepository;
+    this.graphTable = graphTable;
+
+    this.gitMacheteRepositoryFactory = RuntimeBinding.instantiateSoleImplementingClass(IGitMacheteRepositoryFactory.class);
+    this.branchLayoutManagerFactory = RuntimeBinding.instantiateSoleImplementingClass(IBranchLayoutManagerFactory.class);
+  }
+
+  public static GitMacheteRepositoryUpdateTask of(Project project, GitRepository gitRepository, IGraphTable graphTable) {
+    return new GitMacheteRepositoryUpdateTask(project, gitRepository, graphTable);
+  }
+
+  @Override
+  public void run(ProgressIndicator indicator) {
+    // We can't queue repository update (onto a non-UI thread) and graph table refresh (onto the UI thread) separately
+    // since those two actions happen on two separate threads
+    // and graph table refresh can only start once repository update is complete.
+
+    // Thus, we synchronously run repository update first...
+    Option<IGitMacheteRepository> gitMacheteRepository = updateRepository();
+
+    // ... and only once it completes, we queue graph table update onto the UI thread.
+    LOG.debug("Queuing graph table refresh onto the UI thread");
+    // A bit of a shortcut: we're accessing filesystem even though we may be on UI thread here;
+    // this shouldn't ever be a heavyweight operation, however.
+    Path macheteFilePath = getMacheteFilePath(gitRepository);
+    boolean isMacheteFilePresent = Files.isRegularFile(macheteFilePath);
+
+    GuiUtils.invokeLaterIfNeeded(
+        () -> graphTable.refreshModel(gitMacheteRepository.getOrNull(), macheteFilePath, isMacheteFilePresent),
+        NON_MODAL);
+  }
+
+  /**
+   * Updates repository which is the base of graph table model. The change will be seen after
+   * {@link GitMacheteGraphTable#refreshModel} completes.
+   */
+  private Option<IGitMacheteRepository> updateRepository() {
+    Path mainDirectoryPath = getMainDirectoryPath(gitRepository);
+    Path gitDirectoryPath = getGitDirectoryPath(gitRepository);
+    Path macheteFilePath = getMacheteFilePath(gitRepository);
+    boolean isMacheteFilePresent = Files.isRegularFile(macheteFilePath);
+
+    LOG.debug(() -> "Entering: mainDirectoryPath = ${mainDirectoryPath}, gitDirectoryPath = ${gitDirectoryPath}" +
+        "isMacheteFilePresent = ${isMacheteFilePresent}");
+    IBranchLayoutManager branchLayoutManager = branchLayoutManagerFactory.create(macheteFilePath);
+    graphTable.setBranchLayoutWriter(branchLayoutManager.getWriter());
+    if (isMacheteFilePresent) {
+      LOG.debug("Machete file is present. Try to create GitMacheteRepository instance");
+
+      return Try.of(() -> {
+        IBranchLayout branchLayout = createBranchLayout(branchLayoutManager);
+        return gitMacheteRepositoryFactory.create(mainDirectoryPath, gitDirectoryPath, branchLayout);
+      }).onFailure(this::handleUpdateRepositoryExceptions).toOption();
+    } else {
+      LOG.debug("Machete file is absent. Setting repository reference to null");
+      return Option.none();
+    }
+  }
+
+  private void handleUpdateRepositoryExceptions(Throwable t) {
+    LOG.error("Unable to create Git Machete repository", t);
+
+    // Getting the innermost exception since it's usually the primary cause that gives most valuable message
+    Throwable cause = t;
+    while (cause.getCause() != null) {
+      cause = cause.getCause();
+    }
+    String exceptionMessage = cause.getMessage();
+
+    VcsNotifier.getInstance(project).notifyError("Repository instantiation failed",
+        exceptionMessage != null ? exceptionMessage : "");
+
+    IntelliJLoggingUtils.showErrorDialog(exceptionMessage != null
+        ? exceptionMessage
+        : "Repository instantiation failed. For more information, please look at the IntelliJ logs");
+  }
+
+  private IBranchLayout createBranchLayout(IBranchLayoutManager branchLayoutManager) throws MacheteFileReaderException {
+    return Try.of(() -> branchLayoutManager.getReader().read())
+        .getOrElseThrow(e -> {
+          Option<@Positive Integer> errorLine = ((BranchLayoutException) e).getErrorLine();
+          return new MacheteFileReaderException("Error occurred while parsing machete file" +
+              (errorLine.isDefined() ? " in line ${errorLine.get()}" : ""), e);
+        });
+  }
+
+}
